Moved the DSPowSolution to MiningData
