+++ src/libData/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory (AccountData)
 add_subdirectory (BlockData)
+add_subdirectory (MiningData)
