+++ src/libData/MiningData/DSPowSolution.cpp
