+++ src/libData/MiningData/DSPowSolution.h
